requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
validators==0.19.0
xlrd==2.0.1
